requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
colorama==0.4.6
distro==1.9.0
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.68.2
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.0
